Added tof analysis for prudaq
