+++ src/firmware/prudaq/Makefile
@@ -1,7 +1,7 @@
 # can run `make Q=` to see commands as they run
 Q := @
 
-CFLAGS += --std=gnu99 -O2 -Wall
+CFLAGS += --std=gnu99 -O2 -Wall -lzmq
 
 CC := $(Q)$(CC)
 RM := $(Q)$(RM)
@@ -10,7 +10,7 @@ DTC := $(Q)dtc
 
 .PHONY: all clean install
 
-TARGETS := prudaq_capture pru0.bin pru1.bin prudaq-00A0.dtbo
+TARGETS := tof_analysis prudaq_capture pru0.bin pru1.bin prudaq-00A0.dtbo
 
 all: $(TARGETS)
